# internal/model/event.go
package model

import "time"

type Event struct {
	ID          int       `json:"id" db:"id"`
	Title       string    `json:"title" db:"title"`
	Description string    `json:"description" db:"description"`
	Location    string    `json:"location" db:"location"`
	Date        time.Time `json:"date" db:"date"`
	Capacity    int       `json:"capacity" db:"capacity"`
	CreatorID   int       `json:"creator_id" db:"creator_id"`
	CreatedAt   time.Time `json:"created_at" db:"created_at"`
	UpdatedAt   time.Time `json:"updated_at" db:"updated_at"`
}

type EventCreate struct {
	Title       string    `json:"title" binding:"required"`
	Description string    `json:"description" binding:"required"`
	Location    string    `json:"location" binding:"required"`
	Date        time.Time `json:"date" binding:"required"`
	Capacity    int       `json:"capacity" binding:"required,min=1"`
}

type EventUpdate struct {
	Title       string    `json:"title"`
	Description string    `json:"description"`
	Location    string    `json:"location"`
	Date        time.Time `json:"date"`
	Capacity    int       `json:"capacity" binding:"omitempty,min=1"`
}